Formatting and code modifications requested completed
