+++ features/flows/elastic/src/main/java/org/opennms/netmgt/flows/elastic/ElasticFlowRepository.java
@@ -117,17 +117,17 @@ public class ElasticFlowRepository implements FlowRepository {
     
     private String eventIndexName = "";
 
-	public void setEventIndexName(String eventIndexName) {
-		Objects.requireNonNull(eventIndexName);
-		this.eventIndexName = eventIndexName;
-	}
+    public void setEventIndexName(String eventIndexName) {
+        Objects.requireNonNull(eventIndexName);
+        this.eventIndexName = eventIndexName;
+    }
 
-	public String getModifiedIndex() {
-		if (!eventIndexName.isEmpty()) {
-			return String.format("%s-%s", eventIndexName, TYPE);
-		}
-		return TYPE;
-	}
+    private String getModifiedIndex() {
+        if (!eventIndexName.isEmpty()) {
+            return String.format("%s-%s", eventIndexName, TYPE);
+        }
+        return TYPE;
+    }
 
     /**
      * Flows/second throughput
@@ -190,7 +190,9 @@ public class ElasticFlowRepository implements FlowRepository {
         this.nodeDao = Objects.requireNonNull(nodeDao);
         this.snmpInterfaceDao = Objects.requireNonNull(snmpInterfaceDao);
         this.bulkRetryCount = bulkRetryCount;
-		this.indexSelector = new IndexSelector(getModifiedIndex(), indexStrategy, maxFlowDurationMs);
+        this.indexSelector = new IndexSelector(getModifiedIndex(),
+                                               indexStrategy,
+                                               maxFlowDurationMs);
         this.identity = identity;
         this.tracerRegistry = tracerRegistry;
 
@@ -250,8 +252,8 @@ public class ElasticFlowRepository implements FlowRepository {
                 final Bulk.Builder bulkBuilder = new Bulk.Builder();
                 for (FlowDocument flowDocument : documents) {
                 	
-					final String index = indexStrategy.getIndex(getModifiedIndex(),
-							Instant.ofEpochMilli(flowDocument.getTimestamp()));
+                                                                                    final String index = indexStrategy.getIndex(getModifiedIndex(),
+                                                                                                                                Instant.ofEpochMilli(flowDocument.getTimestamp()));
                     final Index.Builder indexBuilder = new Index.Builder(flowDocument)
                             .index(index)
                             .type(TYPE);
