+++ features/jest/client/src/main/java/org/opennms/plugins/elasticsearch/rest/template/DefaultTemplateLoader.java
@@ -54,8 +54,9 @@ public class DefaultTemplateLoader implements TemplateLoader {
         for (int i = serverVersion.getMajor(); i >= 0; i--) {
             final String versionSuffix = i == 0 ? "" : String.format(".es%d", i);
             final String resourceWithSuffix = String.format("%s%s.json", resource, versionSuffix);
-            LOG.debug("Attempting to find template with resource name: {} (requested: {})", resourceWithSuffix, resource);
-			final String template = getTemplate(resourceWithSuffix);
+            LOG.debug("Attempting to find template with resource name: {} (requested: {})",
+                      resourceWithSuffix, resource);
+            final String template = getTemplate(resourceWithSuffix);
             if (template != null) {
                 LOG.info("Using template with resource name: {} (requested: {})", resourceWithSuffix, resource);
                 return template;
@@ -66,13 +67,13 @@ public class DefaultTemplateLoader implements TemplateLoader {
                 serverVersion, resource));
     }
     
-	protected String getTemplate(String resource) throws IOException {
-		Optional<Path> configPath = ConfigFileConstants.getConfigFilePathByName(resource);
-		if (configPath.isPresent()) {
-			return new String(Files.readAllBytes(configPath.get()));
-		}
-		return getResource(resource);
-	}
+    protected String getTemplate(String resource) throws IOException {
+        Optional<Path> configPath = ConfigFileConstants.getConfigFilePathByName(resource);
+        if (configPath.isPresent()) {
+            return new String(Files.readAllBytes(configPath.get()));
+        }
+        return getResource(resource);
+    }
 
 	protected String getResource(String resource) throws IOException {
 		try (InputStream inputStream = getResourceAsStream(resource)) {
