+++ features/opennms-es-rest/pom.xml
@@ -147,10 +147,6 @@
       <version>${jestGsonVersion}</version>
       <scope>provided</scope>
     </dependency>
-    <dependency>
-	<groupId>org.opennms.core</groupId>
-	<artifactId>org.opennms.core.lib</artifactId>
-</dependency>
 
     <!-- Test Scope -->
     <!--
