+++ features/situation-feedback/elastic/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -38,7 +38,7 @@
             <cm:property name="settings.index.number_of_replicas" value="" />
             <cm:property name="settings.index.refresh_interval" value="" />
             <cm:property name="settings.index.routing_partition_size" value="" />
-            <cm:property name="eventIndexName" value="" />
+	<cm:property name="eventIndexName" value="" />
         </cm:default-properties>
     </cm:property-placeholder>
